Rename package to integritybackend to match repo name. (#81)
