+++ integritybackend/asset_helper.py
