+++ integritybackend/exif.py
