+++ tests/conftest.py
@@ -5,7 +5,7 @@ import pytest
 def reverse_geocode_mocker(mocker):
     def patch(return_value):
         mocker.patch(
-            "starlingcaptureapi.geocoder.Geocoder.reverse_geocode",
+            "integritybackend.geocoder.Geocoder.reverse_geocode",
             lambda *_: return_value,
         )
